+++ resources/dynamic/Initializer.js
@@ -18,8 +18,11 @@ dim numberOfMonths
 dim mainCalendar
 dim minDate
 dim maxDate
+dim useScript
 dim minBound
 dim maxBound
+dim minScript
+dim maxScript
 %}
 (function () {
     var adcdefault = new AdcDefault({
@@ -36,6 +39,9 @@ dim maxBound
         hideOrDisable : {%= On((CurrentADC.PropValue("hideOrDisable")) = "disable",true,false) %},
         deciSeperator : '{%:= Interview.Language.DecimalSeparator %}'
     });
+{% If (column.Type = "single" or column.Type = "multiple" or column.Type = "single-loop" or column.Type = "multiple-loop") and CurrentADC.PropValue("useList") = "1" Then %}
+    AskiaSelect2Loader.init('#adc_{%= CurrentADC.InstanceId %} select.askia-select2');
+{% EndIf %}
 {% If (column.Type = "datetime") and Not(column.IsDateOnly) Then %}
        var timePickerR1C1 = new TimePicker({
            showSeconds: {%= CurrentADC.PropValue("showSeconds") %},
@@ -65,6 +71,7 @@ dim maxBound
 	showMonthAfterYear = CurrentADC.PropValue("showMonthAfterYear")
 	numberOfMonths = CurrentADC.PropValue("numberOfMonths")
 	mainCalendar = "'"+CurrentADC.PropValue("mainCalendar")+"'"
+	useScript = CurrentADC.PropValue("useScript").ToNumber()
 	minDate = column.MinDate.Format("yyyy-MM-dd")
 	maxDate = column.MaxDate.Format("yyyy-MM-dd")
 	minBound = column.MinDate.Format("yyyy").ToNumber()
@@ -77,6 +84,8 @@ dim maxBound
 	if CvDkNa(maxBound) < 1 Then
 		maxBound = 2100
 	EndIf
+	
+
 
 	inputName = column.InputName()
 	inputId     = (inputName + "_" + 1).Replace("D", "askia-input-dateO") %}